Fix premature combo firing at 16-bit timer wraparound
timer_read() is uint16_t but last_modify_time is uint32_t, so the elapsed
subtraction was evaluated in 32-bit arithmetic. After the ~65 s timer wrap,
timer_read() < last_modify_time produced a huge positive value that always
exceeded COMBO_WAIT_TIME, firing the combo early. Cast the difference to
uint16_t so wraparound is handled correctly.

+++ src.c
@@ -396,7 +396,7 @@ void combo_user_timer(void) {
   for (int i = 0; i < combo_stack_size; ++i) {
     Combo* combo = &combo_stack[i];
     if (combo->state == 1) {
-      if (timer_read() - combo->last_modify_time > COMBO_WAIT_TIME) {
+      if ((uint16_t)(timer_read() - combo->last_modify_time) > COMBO_WAIT_TIME) {
         ComboPos pos = combo_get_pos(combo);
         if (neq_combo_pos(pos, NONE_COMBO_POS)) {
           combo_press(pos, true);
@@ -405,7 +405,7 @@ void combo_user_timer(void) {
         }
       }
     } else if (combo->state == 4) {
-      if (timer_read() - combo->last_modify_time > COMBO_WAIT_TIME) {
+      if ((uint16_t)(timer_read() - combo->last_modify_time) > COMBO_WAIT_TIME) {
         ComboPos pos = combo_get_pos(combo);
         if (neq_combo_pos(pos, NONE_COMBO_POS)) {
           combo->state = 2;
